backend/pyproject.toml
[project]
name = "self-correcting-agent"
version = "0.1.0"
description = "Self-correcting code generation agent with drift detection"
requires-python = ">=3.13"
dependencies = [
    # FastAPI & Server
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "python-dotenv==1.0.1",
    "pydantic==2.10.4",
    "python-multipart==0.0.20",
    # LLM Providers
    "ollama>=0.4.0",
    "groq==0.15.0",
    # Orchestration
    "langgraph==0.2.62",
    "langchain-core==0.3.29",
    # Evaluation & Metrics
    "datasets==3.2.0",
    "scipy==1.15.0",
    "wandb==0.19.1",
]

[project.scripts]
serve = "uvicorn main:app --host 0.0.0.0 --port 8000"

[tool.uv]
dev-dependencies = []
